Move all API related steps to separate class
